SetPlaylist/requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.3
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
